moved apex functionality to `apex.py`
